Add Interval Matrixes to the Tuning Display
closes #2614
